+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * New cop `Rails/EnumUniqueness` checks for duplicate values defined in enum config. ([@olliebennett][])
 * New cop `Rails/EnumUniqueness` checks for duplicate values defined in enum config hash. ([@olliebennett][])
 * [#3451](https://github.com/bbatsov/rubocop/issues/3451): Add new `require_parentheses_when_complex` style to `Style/TernaryParentheses` cop. ([@swcraig][])
+* [#3600](https://github.com/bbatsov/rubocop/issues/3600): Add new `Bundler/OrderedGems` cop. ([@tdeo][])
 
 ### Changes
 
@@ -2505,3 +2506,4 @@
 [@aroben]: https://github.com/aroben
 [@olliebennett]: https://github.com/olliebennett
 [@aesthetikx]: https://github.com/aesthetikx
+[@tdeo]: https://github.com/tdeo